+++ mac-robber.py
@@ -1,8 +1,8 @@
 #!/usr/bin/env python
 #
 # Author:  Jim Clausing
-# Date:    2024-12-30
-# Version: 1.5.1
+# Date:    2025-02-13
+# Version: 1.5.2
 #
 # Desc: rewrite of the sleithkit mac-robber in Python
 # Unlinke the TSK version, this one can actually includes the MD5 & inode number
@@ -40,7 +40,7 @@ except ImportError or ModuleNotFoundError:
 else:
     have_statx = 1
 
-__version_info__ = (1, 5, 1)
+__version_info__ = (1, 5, 2)
 __version__ = ".".join(map(str, __version_info__))
 
 
@@ -87,6 +87,10 @@ def mode_to_string(mode):
 def process_item(dirpath, item):
     md5 = hashlib.md5()
     fname = os.path.join(dirpath, item)
+    if os.path.islink(fname) and status.st_size > 0:
+        fname_out = fname + " -> " + os.readlink(fname)
+    else:
+        fname_out = fname
     if args.exclude and (fname in args.exclude or dirpath in args.exclude):
         return
     try:
@@ -117,8 +121,6 @@ def process_item(dirpath, item):
     else:
         md5str = "0"
     mode = mode_to_string(status.st_mode)
-    if os.path.islink(fname) and status.st_size > 0:
-        mode = mode + " -> " + os.readlink(fname)
     if sys.version_info < (2, 7, 0):
         mtime = "%17.6f" % (status.st_mtime)
         atime = "%17.6f" % (status.st_atime)
@@ -149,7 +151,7 @@ def process_item(dirpath, item):
     return (
         md5str
         + "|"
-        + fname
+        + fname_out
         + "|"
         + str(inode)
         + "|"